Apply fixes from Benoit Leblanc to address int/double type warnings.

+++ include/openbabel/conformersearch.h
@@ -369,10 +369,10 @@ namespace OpenBabel {
       void SetNbNiches (int value) {nb_niches = value;}
       
       /* @brief Get niches radius, for dynamic niche sharing.*/
-      int GetNicheRadius () {return niche_radius;}
+      double GetNicheRadius () {return niche_radius;}
       
       /* @brief Set niches radius, for dynamic niche sharing.*/
-      void SetNicheRadius (int value) {niche_radius = value;}
+      void SetNicheRadius (double value) {niche_radius = value;}
       
       /* @brief Get the alpha sharing parameter */
       double GetAlphaSharing () {return alpha_share;}
@@ -448,9 +448,9 @@ namespace OpenBabel {
       OBRandom unique_generator; //!< A unique random number generator for the whole algo
       bool use_sharing;		//!< Wether to use sharing or not.
       double alpha_share;	//!< The alpha parameter in sharing function
-      int sigma_share;		//!< The sigma parameter in sharing function
+      double sigma_share;		//!< The sigma parameter in sharing function
       int nb_niches;		//!< The number of dynamic niches to be found
-      int niche_radius;		//!< A pre-determined niche radius, for dynamic niche sharing.
+      double niche_radius;		//!< A pre-determined niche radius, for dynamic niche sharing.
       double p_crossover;	//!< Crossover probability
       double niche_mating;	//!< Probability of forcing the second parent in the first parent
       int local_opt_rate;       //!< Perform a random local optimization every local_opt_rate generations. Disabled if set to
